# backend/docs/swagger.yaml
definitions:
...
  dto.CommentAuthor:
    properties:
      avatar_url:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  dto.CommentResponse:
    properties:
      author:
        $ref: '#/definitions/dto.CommentAuthor'
      content:
        type: string
      created_at:
        type: string
      edited:
        type: boolean
      id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateCommentRequest:
    properties:
      content:
        maxLength: 2000
        minLength: 1
        type: string
    required:
    - content
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.PagedCommentResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.CommentResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  dto.UpdateCommentRequest:
    properties:
      content:
        maxLength: 2000
        minLength: 1
        type: string
    required:
    - content
    type: object
...
info:
  contact: {}
paths:
...
  /mangas/{mangaID}/chapters/{chapterID}/comments:
    get:
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      - description: Chapter ID
        in: path
        name: chapterID
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PagedCommentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List chapter comments
      tags:
      - comment
    post:
      consumes:
      - application/json
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      - description: Chapter ID
        in: path
        name: chapterID
        required: true
        type: string
      - description: Comment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Post a chapter comment
      tags:
      - comment
...
  /mangas/{mangaID}/comments:
    get:
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PagedCommentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List manga-level comments
      tags:
      - comment
    post:
      consumes:
      - application/json
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      - description: Comment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Post a manga-level comment
      tags:
      - comment
  /mangas/{mangaID}/comments/{commentID}:
    delete:
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a comment (comment owner or manga owner)
      tags:
      - comment
    patch:
      consumes:
      - application/json
      parameters:
      - description: Manga ID
        in: path
        name: mangaID
        required: true
        type: string
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: string
      - description: Updated content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Edit a comment (owner only)
      tags:
      - comment
...
swagger: "2.0"